+++ search/searcher/search_filter.go
@@ -53,7 +53,7 @@ func (f *FilteringSearcher) Next(ctx *search.Context) (*search.DocumentMatch, er
 	return nil, err
 }
 
-func (f *FilteringSearcher) Advance(ctx *search.Context, number int) (*search.DocumentMatch, error) {
+func (f *FilteringSearcher) Advance(ctx *search.Context, number uint64) (*search.DocumentMatch, error) {
 	adv, err := f.child.Advance(ctx, number)
 	if err != nil {
 		return nil, err
